+++ backend/src/schemas/card.schema.ts
@@ -4,7 +4,7 @@ export const createCardSchema = z.object({
   title: z.string().min(1, 'Title is required').max(255),
   description: z.string().max(1000).optional(),
   columnId: z.string().uuid('Invalid column ID'),
-  position: z.number().int().optional(),
+  position: z.number().optional(),
   targetIndex: z.number().int().optional(),
 });
